Add TripleO CI scenarios to Puppet gate
We tested TripleO CI scenario001 and scenario004 in puppet-ceph gate and
it works fine. This patch completes the work with other modules:

- Add scenario00(2,3)-multinode-puppet jobs in
  jenkins/jobs/projects.yaml.
- Create Zuul Layout for tripleo-scenario-jobs-scenario00(2,3) groups of
  jobs.
- tripleo scenario001 will be run for puppet-aodh, puppet-ceilometer,
  puppet-gnocchi and puppet-cinder.
- tripleo scenario002 will be run for puppet-barbican, puppet-swift,
  puppet-zaqar and puppet-cinder.
- tripleo scenario003 will be run for puppet-sahara, puppet-mistral and
  puppet-horizon.
- puppet-aodh, puppet-ceilometer, puppet-cinder, won't run
  tripleo-puppet-jobs layout anymore but it's on purpose.
- puppet-cinder will run tripleo ovb jobs for newton and mitaka only
  and scenario001 and 002 in ocata.

This patch is part of the efforts made in TripleO CI to have scenarios,
like we already have in Puppet CI. We also made our best to build
similar scenarios between TripleO and Puppet CIs, to bring consistency
in our tests.

Change-Id: I9af581195d1872cb5fd6f06802d6b8940e66645e