update timeouts for some trove jobs
recently with the increased number of tests being conducted in the
functional and each of the scenario tests, we've been seeing a large
number of timeouts. further, the timeouts seem to only occur in the
rax cloud, things go fine in the osc cloud.

A review of the failures indicate that we're often within 10 minutes
of finishing when the job is terminated. However, in some cases we
would be approximately 15 or so minutes away.

The intent of this commit is to push the timeouts up with some
possible padding and I'll back them off once we know how the system
works with the xenial jobs.

We have done other things to reduce the amount of time to run tests
like splitting some jobs in two (see the single and multi jobs in
scenario as an example).

I know that with larger machines the tests run much faster, we
(tesora) operate a CI in rax and we don't take nearly as long in those
tests. So, given the limitation of the infrastructure in the upstream
CI, and given that we will be adding more tests given the number of
new features going into the release(s), this appears to be the only
solution in the short term.

Change-Id: I7fb49b4e95d06c58cf5bb7efb4a8961dcc6eb629