Switch to Xenial: mistral-docker-buildimage
Note that this job produces a single output, a docker image:
http://tarballs.openstack.org/mistral/images/

It is the same image for all branches. Therefore run the job only on
master.

The job is in post and experimental queue, remove voting, it's not used
for these.

Change job to xenial as part of transition of master jobs to run on
xenial.

Change-Id: I011df56ee059d67c54e08b6e1a32faaa87d071c9