Remove multinode grenade in Cinder's stable/mitaka
A Cinder live upgrade from Liberty to Mitaka should work, but we weren't
gating on that back in Mitaka timeframe and the feature was considered
experimental. There's no need to gate on that in stable/mitaka. Moreover
the job is currently constanly failing due to PyMSQL not being installed
on the subnode.

This commit disables the job for Cinder's stable/mitaka and unblocks
the gate.

Closes-Bug: 1647672
Change-Id: I362d1306d201f5f239b228d6df2108349517295e