Switch to Xenial: zaqar
Convert zaqar jobs to xenial. Newton and newer will run on xenial,
older branches on trusty.

Let the gate-dsvm-*zaqar run only on trusty and the newer
functional/tempest ones on xenial - following the intention of the
comments.

Mark the gate-dsvm-*zaqar as -nv instead of using special condition to
make them non-voting.

Change-Id: I8654dbcf02d0c8cf5cc91d993bc9a35010d5758a