Convert Designate jobs to Xenial
This is part of the Trusty to Xenial transition. For changes against
branches >= newton run on Xenial. Older branches continue to run on
Trusty.

Note that this updates jobs to use use the pdns4 driver on Xenial as the
powerdns driver is for older versions of powerdns.

Also this updates the designate grenade jobs which have the following
layout:

  Trusty:
    mitaka -> newton
  Xenial
    newton -> master/ocata

Change-Id: I1fc6c629364336418733fcf2564c2c28b5589dcb