Finish heat transition to Xenial
This cleans up a small number of straggler heat jobs to run on Xenial.
This affect changes to branches >= newton. Jobs against older branches
continue to run on Trusty.

This also updates the identity v3 only jobs to run against xenial
beacuse this job wasn't passing until recently there is little reason to
run it on trusty and older branches.

Change-Id: Ieff5f25c672c65f3e3415e4d98b6d7e2c24d41d4