Xenialify Java builds
This is part of the transition to Xenial. This updates monasca and
clouddocs java builds to run on xenial for branches >= newton. Older
branches still run on Trusty.

This does not change the gerrit and jenkins plugin build which continue
on Trusty because that is where we deploy these services.

Change-Id: I778dde7e24d220a892ba045d650920d28bec24d0