Disable broken horizon integration tests
The PTL of horizon asked me to help make these jobs non-voting
as they are currently not passing. Further these jobs are already
non-voting against master.

This is a regression against the stable branches which in all reality
is unlikely to get the attention of developers to ever turn it back
on. I have not investigate why the job is no longer functional.

For those reasons I'd like the PTL of stable and PTL of Horizon to
approve this change.

Change-Id: I6522a98cfca321c18be56f1970ba51098fd1b58d