Move nodepool integration testing to Xenial
Since devstack and the rest of OpenStack are moving testing for master
ahead to Xenial we should follow along with nodepool integration
testing. This has the drawback of not testing nodepool on the same
distro release we currently deploy on but that seems to be a small risk
compared to deploying an entire cloud on distro it isn't tested on.

Nodepool compared to OpenStack is relatively simple and should be happy
on Trusty or Xenial both.

Change-Id: Id00a069994e9d4443dec483332dbd3787a61fe10