neutron: switch experimental centos job to single node
As was revealed in comments to
I8d63a883b29eda2757656008be3c6af83b8f9e88, CentOS is not stable for
multinode gate setups just yet, so an attempt to make it work for
Neutron is probably too big a bite for us.

Instead, let's scale back and try a single node, no DVR job and see if
we can make it work. If we suceed with it, we can look at
expanding/replacing it with a dvr flavour.

Change-Id: Ida69e44c38fc2f91eaff121b3078004d8b376808