Handle cinder driverfixes branch (part1)
The driverfixes/* branches will be used to provide a place for vendor
bug fix backports that do not meet stable policy, and as such testing
for these patches will be on the driver vendors to ensure they should be
made available to their customers. This filters defined Cinder jobs to
only run pep8 and releasenotes tests against these branches.

The releasenotes test is only run if releasenotes are changed and should
stay to not get broken releasenotes in.

Part 1: Handle all jobs that are specific to cinder.

These jobs are run in check/gate queue for cinder and are handled
explicitly by this change:

* gate-cinder-pep8-ubuntu-trusty
* gate-cinder-pep8-ubuntu-xenial
* gate-cinder-docs-ubuntu-xenial
* gate-cinder-pep8-ubuntu-xenial
* gate-cinder-python27-db-ubuntu-xenial
* gate-cinder-python34-db
* gate-cinder-python35-db
* gate-cinder-requirements
* gate-cinder-tox-compliance-ubuntu-xenial
* cinder-* post jobs

Changes like hese are ubuntu-trusty-jobs that will only run on
stable/mitaka and stable/liberty, thus never on driverfixes:
* gate-cinder-python27-db-ubuntu-trusty
* gate-grenade-dsvm-neutron-ubuntu-trusty
* gate-tempest-dsvm-full-devstack-plugin-ceph-ubuntu-trusty
* gate-grenade-dsvm-cinder-multinode-ubuntu-trusty
* gate-tempest-dsvm-neutron-full-ubuntu-trusty
* gate-rally-dsvm-cinder-ubuntu-trusty-nv
* etc.

Handled elsewhere:
* gate-cinder-api-ref (only runs on master)

These changes are not yet handled:
* gate-tempest-dsvm-neutron-full-ubuntu-xenial
* gate-grenade-dsvm-neutron-ubuntu-xenial
* gate-tempest-dsvm-full-devstack-plugin-ceph-ubuntu-xenial
* gate-grenade-dsvm-cinder-multinode-ubuntu-xenial
* gate-rally-dsvm-cinder-ubuntu-xenial-nv
* gate-cinder-dsvm-apache-ubuntu-xenial-nv
* gate-tempest-dsvm-full-sheepdog-ubuntu-xenial-nv
* gate-tempest-dsvm-full-drbd-devstack-ubuntu-xenial-nv
* gate-tempest-dsvm-full-lio-ubuntu-xenial
* gate-tempest-dsvm-full-bdd-nv
* gate-tempest-dsvm-neutron-identity-v3-only-full-ubuntu-xenial-nv

Change-Id: Ia88441b437efeda388c00678ebcc72bca15d45fb
Co-Authored-By: Sean McGinnis <sean_mcginnis@dell.com>