Split magnum shade support into its own job
The magnum team reports problems with having magnum and all the other
services enabled at the same time due to memory pressure. We see a
flakier gate when we have magnum enabled than when we don't. So - split
out magnum support into its own job that does not include swift, heat or
designate, then include swift, heat and designate in the other jobs.
Finally, turn off horizon and ceilometer from the gate since we do not
test either of them.

Change-Id: I7e7ede279be2acabd469fbe997ae54743768dfd8