Make ...ironic-pxe_ipmitool-pyghmi-src-ubuntu-xenial voting
This patch is making the
"gate-tempest-dsvm-ironic-pxe_ipmitool-pyghmi-src-ubuntu-xenial" voting.

Ironic is becoming more dependent on pyghmi because it relies on the
pyghmi code to test the dpeloyment of the nodes using the pxe_ipmitool
driver in gate, so, let's make this job voting to avoid changes in
pyghmi to break Ironic.

Change-Id: I671e281a875b1a109c6a8dcdda37bec65002bca4