[Manila] Disable tempest for grenade job
Disable tempest for manila grenade job in proper way.
We have not installed it intentionally before and haven't added
to list of enabled services. But it tries to be updated in grenade
job. And after merge of [1] change to grenade project, that update
started failing. So, disable upgrade by specifying appropriate env vars.

[1] Idede322190d57491d4f7e18090aa8c29b5b87353

Change-Id: I1573df1fcd35564d7240efca2ec39304eb54ea60
Closes-Bug: #1650482