Remove z/tempest
Change I3749f88edd1984286f9cfc840bfe37722b2f28be moved the comment for
z/tempest away from it. Remove it.

Remove z/tempest completely, with just a single job in it, it's useless
- and not needed anymore.

The current job to name the integrated queue is for a newton only
change, use another job from the integrated gate to name the queue.

Change-Id: I138c7c39d0dcf2e65f00f34cbb160ab0d5b49813