Fix zuul layout order for tripleo scenarios
Zuul layout order wasn't correct for tripleo scenarios, the jobs were
still running on stable/hammer branch because the rule was overriden by
the rule moved in this patch.

Moving this 'general' rule before specific rules makes the layout safer
so we first define the general rule and then the specific rules for the rest.

Change-Id: I77c342d57302c0e6a0d550e3a211b06190fd4666