devstack: enable tls-proxy for all -platform- jobs
Ubuntu platforms are known to work just fine with it, as can be proved
by other jobs having it enabled and passing in the gate.

As for CentOS flavour of the job, there were some concerns around some
issues with devstack deployment with it being enabled. I just checked it
locally on a 7.3 machine, and it passes ./stack.sh just fine with
'enable_service tls-proxy', so it's not clear which issues we envision.

I suggest we enable the service for all -platform- jobs and see what
breaks for centos. If it still works just fine, we can proceed moving
that job into integrated gate. If it fails, we will deal with it, and
it's not the end of the world, because the jobs are all non-voting.

Change-Id: I322b24e69271667b0c27c8edbba1a0bcea02c525