networking-odl: new experimental gate job for new netvirt
Added a new experimental gate non-voting job to test on new netvirt of
OpenDaylight.

The reason why we use a new job template to do this is that it will make
it much more easy for us to switch the whole CI to new netvirt without
changing many of the CI job names.

Change-Id: I81c8a97ebfdbe38fcfd86a9d8dd7e7ce9c3e224b
Partial-Bug: #1643948
Partial-Bug: #1646940