Use Ironic Inspector tempest plugin from master
Fetch master to install the plugin if a patch isn't against
master. And on master, use the local change to pick up any
changes to the plugin.
Ironic is doing this already.

Otherwise it could have side effect when tempest breaks Inspector
grenade, where we need to fix tempest plugin for both branches.
And it's hard to do with voting grenade job
I2660fda0c5aeca656be5c7b565f2e0bf255658f1.

Also fix ironic's version of this to append to TEMPEST_PLUGINS instead
of clobbering it.

Change-Id: I547650f221f481d28931af8e4a45daf17aa3f765