Removing ironic-webclient gating for retirement
Per the annoucement[0] to the openstack-dev mailing list,
this change removes the gating from ironic-webclient to
allow us to retire the ironic-webclient.

Additionally, the '{name}-nodejs{version}-npm-publish-draft'
template was only used by the ironic-webclient to publish a
working draft for users to review. As it is no longer used
anywhere else, it must be removed in order to permit tests
to pass.

[0]: http://lists.openstack.org/pipermail/openstack-dev/2016-December/109195.html

Change-Id: Ie6d940a071f158be647dd86c32752e5988d7c968