Pin only gate mitaka jobs to master
The intent of this rule was to run changes against master of certain
projects against older clouds like mitaka. So we had a rule that said
run all these ubuntu-trusty-mitaka jobs against only master.
Unfortunately I think this prevented the periodic mitaka jobs from
running as they are branchless.

Simple fix is pin to master only when gate- is the job name prefix as
all the periodic jobs start with periodic.

Change-Id: I6eb09d9138e7076c4a93d2de3771204f8a6a4197