Py34/py35 cleanups
Remove branch condition for gate-networking-odl-python34. The global
regex to run only on liberty and mitaka is just fine.

Remove branch conditions for gate-trove-python34-db and
gate-trove-dashboard-python34. Those are against not running on mitaka.
The jobs are not used anymore and python35 is setup.

Remove gate-murano-dashboard-python34 branch condition, the job is not
used.

Remove gate-quark-python34 non-voting condition, the job is not used.

Add py35-nv job for mistral-dashboard, solum-infra-guestagent,
python-tackerclient since they
have a py34 non-voting job but no py35 job.

Change-Id: Ia6d6fe31c33b0e5b007210f099c1c74df52206c4