Teach run-tox-api-ref.sh to install requirements with constraints
In trying to add constraints support to os-api-ref we hit a case where
installing os-api-ref with this script will also update requirements.

If a requirement isn't blocked in thw gate *or* os-api-ref hasn't merged the
proposal-bit's update this can lead to failures.

This change addresses that by installing openstack-requirements and
removing os-api-ref from upper-constraints.txt.  Then when it installs
os-api-ref the version from git is appropriate BUT all requirements are
also constrained.

Note: This assumes running in the gate and that the git setup has copied
the upper-constraints.txt into a local file and set the
UPPER_CONSTRAINTS_FILE appropriately.  The new commands will fail and
cause job failures if these pre-conditions are not met.

Change-Id: Ic9b49dbd2b9b8c991fa6d21545460ce015bcc320
Needed-By: I7ad1c2afda29b095ff3d53db5eec7101f8816c2e