Add check-requirements job to additional tripleo projects
There are a few projects in tripleo that have python dependencies
but aren't currently syncing with g-r.  This adds the jobs to verify
that those projects stay in sync.  The patch to add those projects
to the sync list is If90b61d274748a1250bd9b261a0b42ef93f759ec.

Change-Id: I8d83bc123cf6622c4320f1e981e70515628eca9d