Introduce TripleO multinode (x4) experimental jobs with scenario005
TripleO Architectures allows complex scenarios where we can choose which
services (run with Pacemaker in HA) live on which node.

http://specs.openstack.org/openstack/tripleo-specs/specs/ocata/composable-ha-architecture.html

To test this architecture, we need a 4-nodes deployment where:
- 1 node is for undercloud
- 3 nodes are for Pacemaker cluster with composable services

scenario005 will deploy this complex architecture and makes sure
OpenStack Cloud is working well, like we do with other scenarios.

This patch aims to:
- introduce 4nodes scenario in OpenStack Nodepool.
- introduce scenario005, in experimental pipeline
- configure scenario005 to run only when needed:
  - check experimental
  - only starting from ocata
  - only when we patch pacemaker files in puppet-tripleo and
    tripleo-heat-templates.

Change-Id: Ic1d05ba7d572a291d2c77d85473394564699078c