Remove a few trusty-based jobs from nova experimental
Now that liberty is end of life, stable/mitaka is the only
branch that can run on ubuntu trusty nodes and we want to
pair down our list of available jobs a bit. This patch removes
some trusty-based jobs from the nova experimental queue with the
understanding that nova doesn't really need to run these against
stable/mitaka changes, generally either because they don't provide
good coverage or are necessary for the types of fixes that get
backported to stable/mitaka, or just because they probably don't work
on mitaka anyway.

Change-Id: I8a3ee9ed3df1ed08f987b7ae5ea4bed1ed32e758