Make tripleo ovb updates job voting
This job has been running for a while now and there have not been
any more spurious failures in it than in the other ovb jobs.  Since
it has also caught some legitimate problems in that time but was
ignored because it's non-voting, we'd like to move it to voting.

Change-Id: I3c328699394474d35ad2eb3088d4efb44829c579