Fix kuryr-kubernetes gate
In [0] and [1] we introduced variant install tests for
kuryr-kubernetes.
gate_hook.sh is in the kuryr-kubernetes directory, which is added to
the PROJECTS in the devstackgaterc called by gate_hook.sh - circular.
Add kuryr-kubernetes to PROJECTS before devstack-vm-gate-wrap.sh,
so it will be cloned by ansible, before the gate hook is called.

  [0] https://review.openstack.org/#/c/416170/
  [1] https://review.openstack.org/#/c/416168/

Change-Id: Ifda37db9d811cca270f095c36c18c722b079b8ef