Add releasenotes and python35 jobs to Tricircle
Currently no releasenotes and python35 jobs in Tricircle, the
releasenotes job is basic job to publish a release, and python35
is a common goal for the OpenStack community, so it's time
to add these jobs to Tricircle project.

At the same time, the xenial dvsm job is stable in Tricircle, it's
time to convert it to voting job.

Depends-On: I18cb59cadb7a1c06f6cd729c4bda2c8e95d41e1e

Change-Id: I20936840d0ea73f4b272172cedb0c6b759d6c0cc
Signed-off-by: joehuang <joehuang@huawei.com>