add functional test for osc that uses master libraries
OpenStackClient uses a lot of other openstack libraries, we need
a job that installs these from master and is non-voting so we
can catch failures before they are out in the wild.

Related Patch: Ib6391893b2302bdc514525d5ddda886fe8c60100

The related patch is for adding the underlying structure to
openstackclient, I did not add a Depends-On since the job is
non-voting.

Change-Id: I557955621a66a4746a5923915d13f5d8078bd0e6