Add OSA periodic upgrade job
In order to increase the ability to test various
scenario's for a greenfield deployment and an upgrade
OSA is implementing periodic jobs to achieve this.

These jobs will have a longer timeout to cover
scenarios that are more complex or are essentially
two builds in a row (eg: upgrades).

This is an initial job to test the waters and iron
out the kinks. Once this job is successful we intend
to implement periodic deploy jobs (to guard against
bitrot) and periodic upgrade jobs which test upgrades
from N-2 (currently stable/newton) to N-1 (currently
stable/ocata), and from N-1 to N (master). The jobs
will execute multiple scenarios and on multiple
operating systems.

Co-Authored-By: Jesse Pretorius <jesse.pretorius@rackspace.co.uk>
Change-Id: I63ce88510409f82791d3592b7d2c056e1fe0ff71