Restrict placement job to stable/newton
The gate-tempest-dsvm-neutron-placement-full-ubuntu-xenial-nv job
is only unique for stable/newton since it runs the placement-api
plus cells v2 plus neutron, which are all defaults in master (ocata)
now. The placement-api and cells v2 bits were optional in Newton
so it only makes sense to run this on stable/newton changes now.

Change-Id: I145ed119a7b2ba8548ebbaa8f851fdeb74687320