Move requirements job failure msg into log
A failure on the requirements job currently uses a custom
failure-message.  The problem is that this custom message doesn't
match the gerrit magic for extracting failures and showing them in the
"Jenkins check" box.

This means that jenkins votes -1, but you don't see any failures
unless you go and click the "toggle ci" button.  This leads to
confusion and people rechecking unnecessarily.

The message doesn't seem particlarly important; the linked page is
just the regular documentation and isn't exactly a trouble-shooting
guide.  Thus move the message and link into the job output so this job
remains consistent.  The message stays similar but without the link so
it matches.

Change-Id: I2af14c7993e097178c4cc7ee894804c005148523