Fix restriction for gate-tempest-dsvm(-multinode)?-full-ubuntu-xenial
The regex here was too strict so that it was not matching on the -nv
on the gate-tempest-dsvm-multinode-full-ubuntu-xenial-nv job and was
thus running that job on master (ocata) which is redundant with the
gate-tempest-dsvm-neutron-multinode-full-ubuntu-xenial-nv job since
on ocata+ they are both running neutron now (so the config is the
same).

This loosens the suffix on the regex to capture the non-voting job.

Change-Id: I1c22983f1f1fa09e1ebb6197a541f09f2cf14137