puppet: disable voting on ubuntu/integration jobs
Terrible things happened in Ubuntu lately and we need to patch our
module to work again on Ubuntu.
For that, we need to disable voting for a few hours or days, the time we
fix things and make CI happy again.

This patch will be reverted as soon as we fix things.

Change-Id: Iab46a9f8076c94996677a4834a08e73176e1f54e