Update wheel build job timeout
As seen in [1] it can take longer than 90 minutes to run the wheel
build job.

This is probably due to I70ae19cc974a8c34ad9ff380fee5adbb17c96ff4 and
the job not running for a while.  I will keep an eye on it -- if this
isn't long enough I'll take the effort to run it manually.

[1] http://logs.openstack.org/periodic/wheel-build-centos-7-amd64-python2/0cdc862/console.html

Change-Id: Idc1d53d1e73b4a9eb13f258f4a4d5627ec3cf300