devstack-gate tempest-dsvm-py35 job: don't run c-bak
In Ia486ffcf3868f9ed04d29208e2a58fda3664d895 cinder backup
tests were incorrectly skipped when Swift is not enabled, to
accomodate this tempest-dsvm-py35 job. But
the correct move is to disable the c-bak service if Swift is
not enabled, because in the Gate c-bak can only work with
Swift.

Change-Id: I6b85e6d8cb5833bb0e28dae5e406b3bae30eb36c