Run jobs on mitaka/liberty branches with a suffix
The regexes used to a particular run jobs on < Newton and another
job on other branches results in neither job being added for
branches named 'stable/liberty.+' or 'stable/mitaka.+'. This change
removes the $ from the liberty/mitaka regex to allow for suffixes.

For example Jenkins is not voting on this because the branch name
is stable/mitaka-hpb:

https://review.openstack.org/#/c/424854/

Change-Id: Idf3988604eea43ca247743f8b69e76f237c88e40