change pxc and postgresql scenario tests to nonvoting in trove gate
These tests have become unreliable and unstable and it is now
virtually impossible to get two consistent runs of these tests given
the load on the CI infrastructure. The issue is NOT the CI
infrastructure but the fact that these tests are not
deterministic. Run on a quiet Saturday morning and they run just fine,
run them on a cloud with nested KVM and they run fine most of the
time. Run them when the CI is heavily loaded and they fail
spectacularly. The tests are not ready for prime time as we had hoped
when we made them voting earlier in the Ocata cycle. More of an effort
will be focused on stabilizing these tests in Pike.

Yet, I'm leaving the tests in as non-voting in the check and core
reviewers will look at failures to determine whether they are
introduced by the change or not and make a value judgement on whether
the subject changes should be merged or not. In the remaining short
time left for the Ocata cycle, this is likely the best we can do.

Change-Id: I7a5c008eb47f560cab0c8e0b373e586d3abde726