Revert "Migrate -platform- devstack jobs to -neutron-full-"
This reverts commit b577ae03f173ebb9ec7af84ec56ca6ce67b467b7.

This change unexpectedly added -ubuntu-trusty job in the master
integrated gate (and probably removed it from other branches). What we
need to do is to run it in devstack master AND in integrated gate for
Mitaka. I am not sure how to achieve that, so in the meantime, I revert
the patch.

Change-Id: Icc81edd85abbf8912328145f4c95d784271c4ae0