Enable grenade plugin in Zaqar grenade job
Due to a misunderstanding for the grenade implement, we added the
grenade job into check and gate but we don't have a grenade
plugin in zaqar tree. Now we're adding the grenade plugin by
patch https://review.openstack.org/426943 and we also need to
make the grenade job as non-voting before we're happy with the
test result.

Closes-bug: #1660874

Change-Id: Ie57960ca9ded5efb2e69fa24b2390dc8cf31c46f