OSA: Re-implement func test jobs for stable/mitaka
The intent of https://review.openstack.org/428691 was to
reduce the scope of functional test jobs running against
the stable/mitaka branch, but it reduced the job set to
zero unintentionally. Effectively those jobs are only
run against the stable/newton branch now.

This patch returns the previous include, then remodels
the inclusion before it into an exclusion after it so
that there is only one inclusion for trusty, then follow
on filters exclude the jobs that should not run against
stable/mitaka. This should achieve the goal of only
running the limited job set against mitaka, and also
supporting trusty for newton.

The CentOS job restriction further down in the file is
removed as the same functionality is now consolidated
into the new skip rule by virtue of the fact that it will
skip all non trusty jobs.

Change-Id: Ieb14a7bb49a2c3e5aaa0c169ee179d8087fa4711