Demote trunk gating job to experimental
The trunk gating job is a NOOP and is wasting infrastructure
resources.  We have a very aggresisve experimental->non-voting
promotion process for jobs.  Once the job is completed in our
gating scripts, we will promote it again.  Becaue of our aggressive
promotion process for experimental to gating jobs, experimental is
rarely run and jobs only sit in this queue for about a week.

Change-Id: Idf65c19de814876b84512ae4c66cd9d63699a07a