Kill gate-tempest-dsvm-platform-ubuntu-trusty
In change I6617283afd798af37e64913b7865cea3c8a62aba we are
removing support for Trusty because (1) we don't gate on it
in master and (2) we're raising the minimum required version
of libvirt in Nova to a version that isn't available in
Trusty.

Since the gate-tempest-dsvm-platform-ubuntu-trusty job is
only run on master and is meant to keep the lights on, and
we want to turn the lights off, let's just remove the job.

Change-Id: If69f99bd789e646b0261e27a8a061efde32436f7