add tripleo non-voting job into neutron & neutronclient gates
With the motivation of adding more diversity in how do we test
neutron & neutronclient, this patch will add a tripleo job,
non-voting.

The idea is to bring more feedback when testing patches in these
projects and detect what failures we could have outside devstack.

The jobs are not voting. We don't expect Neutron team to open TripleO job
logs. The idea is to let TripleO folks know when a job is red so they
can investigate and tell if something critical is broken in the patch.

Change-Id: Ia60812b7b7ac5cb19d5707ffb5130e804c36a001