Remove Puppet OpenStack beakers jobs on Ubuntu Trusty
Puppet OpenStack beaker jobs on Ubuntu Trusty became unstable for some
reasons we don't really control and we don't have time to fix the jobs.

Though we still have strong testing with puppet-openstack-integration,
which works fine. This patch removes the beaker jobs on Trusty platform
so we don't have to maintain them anymore and keep merging patches in
stable/mitaka and rely on integration jobs that work fine.

Change-Id: Ie6f1206a1b946cd9f89caa9f8c69df734609f2d3