Remove linters job of elastic-recheck
This commit removes the linters job of elastic-recheck. The job was
introduced in I9ee95f177882a0e40018ff0097877eab3e435a2b . However, the
job executions have been failed for a long time. So this commit just
remove it to keep the job clean and to avoid to waste resources.

Change-Id: Ia9534f35885951e884dcb4bcb56c76c6ec4d6a4f