Stop swifts func-post-as-copy job running on stable
The toxenv func-post-as-copy doesn't exist in swift before ocata.
This means any stable patch fails in the gate due to jobs:

  - gate-swift-tox-xfs-tmp-func-post-as-copy-ubuntu-xenial
  - gate-swift-tox-xfs-tmp-func-post-as-copy-ubuntu-trusty

This toxenv was added back in swift via change
4ee20dba482e8f8e4b9ea574428c72c5728a10a1. This change changes
a swift default so can't be backported to stable branches so
instead this change stops these jobs running on older swift
stable branches.

Here is an example error from a stable job:

 + tox -v -efunc-post-as-copy
 using tox.ini:
<snipped>/gate-swift-tox-xfs-tmp-func-post-as-copy-ubuntu-xenial/tox.ini
 ERROR: unknown environment 'func-post-as-copy'

This will unstick the swift stable check and gate pipelines.

In ocata swift now defaults to the fast-post behaviour, meaning
fast-post functional tests happen in the normal func toxenv. But
this means the pre ocata stable branches do not test fast-post anymore
so we've added the func-fast-post job back but only for the releases
that need it.

Co-Authored-By: Tim Burke <tim.burke@gmail.com>

Change-Id: I2bfe8d1c324f3bb2ebba3db47e68c27feac281e1