Add zuulv3 main.yaml
This adds a simple starter zuulv3 main.yaml file for testing the POC
zuul deployment. Note that puppet-zuul copies the entirety of
project-config/zuul into /etc/zuul/layout on the hosts it deploys
zuul to. So this will give us /etc/zuul/layout/main.yaml which we
probably want to change to /etc/zuul/main.yaml in the future.

Change-Id: I028d3da2248ea6a28201d9afe86240b61d6cc772