Re-implement openstack-ansible-tests func job to test ARA
In https://review.openstack.org/#/c/435570/ we added an OSA
job to ARA and later had to revert it because the way the jobs are
designed to work do not fit with ARA.
The current job layout for OSA expects the project being tested
to have a tox env 'functional' which then proceeds to retrieve
openstack-ansible-tests and then fire tests from there.

ARA doesn't have 'functional' tox env but beyond that, we're
interested in testing that openstack-ansible-tests is able to
install and use ARA properly. Running tox from ARA's repository
installs ARA so it's a bit backwards.

Rather than change OSA's current testing methods, we'd much rather
re-implement a slightly changed job that always clone
openstack-ansible-tests and runs the functional tox env from there.
openstack-ansible-tests has the necessary code to zuul-clone ARA if
we're running in the ARA gate.

Change-Id: I883806aab498b7bbd686e5139edec9598f83952f