Skip gate-puppet-openstack-integration-4-.. jobs
gate-puppet-openstack-integration-4-scenario00X-tempest-centos-7-nv
jobs were not skipped even if the patches were not related to the code.
So the jobs wasted the gate resource, so this patch makes them skip when
proposing unrelated patches.
This issue was detected on I6f5c86778e9394ffa56bd11678f47d2939c7ab47

Change-Id: Ic6a7728d6789c37bb57ceea51b295050cd40d242