Install barbican-tempest-plugin through local config
Previously, barbican-tempest-plugin was installed every time the
barbican devstack plugin was enabled, which was breaking things for
other projects that were trying to use barbican in their tests, but
not run the barbican tempest tests themselves (see
Iabbdc5177f62514bf5fc4dff5b79ba3c11083506). This is the first step
to explicitly install the barbican-tempest-plugin where the tests
are wanted.

Change-Id: Ibef3f9a135f14727bf57c29e766f838d7da56c68
Co-Authored-By: Brianna Poulos <Brianna.Poulos@jhuapl.edu>