Remove trove tempest plugin job
The tempest trove plugin only has limited testing that has
duplicated coverage with other testing in trove. There isn't much value
in continuing to use test resources for this. The plugin is also going
to be removed from trove also, so removing the job is the first step for
this.

Needed-By: Iccc9e62bcd46c7aa8b67a454a9882f41f743f332
Change-Id: I0082c483b849deeffe872a5489c7123c2b2c706c