Allow gate hook to distinguish between ovs and lb scenario jobs
This is needed so that we can choose a custom firewall driver (ovsfw)
for openvswitch job flavour. This is needed because one of tests target
connectivity for trunk ports, and this works only when using the new
ovsfw driver.

The gate hook already matches the argument against 'dsvm-scenario'
pattern and doesn't require equality, so this change should not break
existing jobs.

This change relies on I5a36fa3a44e9de4225a00c854a95ad7b5d8987f7.

Depends-On: I5a36fa3a44e9de4225a00c854a95ad7b5d8987f7
Change-Id: Ie3261f528ccedb5fc2b1597c84cf998e82ede2cd