convert dvr-ovs-native job to local_conf stanza
This is one of the projects that used an inline pre_test_hook to set
local.conf. Move it over to the new stanza to test this all works. The
only use of this in the gate is non voting, so we should be able to
land and iterate to make sure that this is really working.

Change-Id: I64f610e963192be48166e8ef638fee85075dcae0