Add barbican-tempest-plugin tests to Castellan
Castellan is used by various projects (nova, glance, cinder, etc.)
for different security features (including image signing, cinder
volume encryption, and ephemeral storage encryption).  As scenario
tests are being added to the barbican-tempest-plugin for these
features, it would be valuable to confirm that changes to castellan
to not break these features.

This change adds an experimental gate check to castellan for
the barbican-tempest-plugin, including setting the variable
LIBS_FROM_GIT=castellan to cause the current patchset of castellan
to be used instead of the latest version from pypi.

Change-Id: Ic5b3e86e8f58916bd645a8a7f0d701e8757830ed