Turn on voting for Nova tox-functional-py35 job
The job has been stable for a while now and we also
have the tests-py3.txt blacklist if we need. So let's
make the job voting from now on. This helps with the
Pike goal as well. Since this job is voting, we need
the job to be in the gate queue as well.

Change-Id: Ib7b8d8eff74f891d257b2ed5b4d35282b32b8694