Drop osic-cloud1-disk flavor from nodepool conf
As part of resource restructuring within OSIC, we're ceasing use of
512 nodes with their "disk" flavor, and continuing to use 256 each
of their "s3500" and "s3700" SSD-backed flavors. Independent of
this, they will be adjusting these flavors from 8vCPU to 4vCPU so
that we can bring our peak utilization in line with their current
availability.

This is expected to be a temporary measure, until they're able to
complete deployment of a dedicated 35-host OpenStack environment
where we can start adjusting overcommit ratios (hopefully driving
our cap back up again somewhat).

Change-Id: I0244362a61264bb36e5c1e1562684a0136944e49