Change TripleO multinode jobs to be quickstart based
We are gradually transitioning the way we run CI in TripleO. We
are now ready to move the multinode jobs to be run via
tripleo-quickstart. This changes the multinode gate and check jobs
run in the check-tripleo-jobs group to be run via quickstart.

Since these jobs will now be in the merge queue for other TripleO
projects they are made voting on tripleo-quickstart/extras as well.

The job types are also changed to reflect the new way we are
parsing them. Eventually we will also change job names to be in
line with these new job types, but we want to start with job names
similar to the current ones in order to cause less confusion.

Change-Id: I8208a4982432dbf8f6bd69d895480bf1baf0ea41