Use gate_hook but not pre_test_hook in Tricircle multiregion job
The multi-region job for Tricircle needs to update some
environment variables which will be used in the devstack based
test environment setup, it's impossible to set these variables
in pre_test_hook.sh without sourcing it, which leads to the job
failure. gate_hook is a better hook point to do this, many other
projects like Octvia, Magnum etc already used this hook point to
configure environment variables. It's better to use gate_hook
but not pre_test_hook in Tricircle multiregion job

Change-Id: I6763d7141895dfc3e94baed92b2c052a698049d1
Signed-off-by: joehuang <joehuang@huawei.com>